+++ forge/src/main/java/io/github/how_bout_no/outvoted/client/render/WildfireHelmetRenderer.java
@@ -2,6 +2,9 @@ package io.github.how_bout_no.outvoted.client.render;
 
 import io.github.how_bout_no.outvoted.client.model.WildfireHelmetModel;
 import io.github.how_bout_no.outvoted.item.WildfireHelmetItem;
+import net.minecraft.client.render.VertexConsumer;
+import net.minecraft.client.render.VertexConsumerProvider;
+import net.minecraft.client.util.math.MatrixStack;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
 import software.bernie.geckolib3.renderers.geo.GeoArmorRenderer;
@@ -19,5 +22,14 @@ public class WildfireHelmetRenderer extends GeoArmorRenderer<WildfireHelmetItem>
         this.leftLegBone = "armorRightLeg";
         this.rightBootBone = "armorLeftBoot";
         this.leftBootBone = "armorRightBoot";
+    }
+
+    @Override
+    public void renderEarly(WildfireHelmetItem animatable, MatrixStack stackIn, float ticks, VertexConsumerProvider renderTypeBuffer, VertexConsumer vertexBuilder, int packedLightIn, int packedOverlayIn, float red, float green, float blue, float partialTicks) {
+        super.renderEarly(animatable, stackIn, ticks, renderTypeBuffer, vertexBuilder, packedLightIn, packedOverlayIn, red, green, blue, partialTicks);
+        if (this.entityLiving.isBaby()) {
+            stackIn.scale(0.7F, 0.7F, 0.7F);
+            stackIn.translate(0F, -0.4F, 0F);
+        }
     }
 }